refactor(vault): decompose VersionHistory into focused sub-components
- Created vault version components:
  - VersionList.tsx (415L) - version list display
  - VersionDiffView.tsx (406L) - diff view component
  - VersionPreviewModal.tsx (176L) - preview modal
  - VersionRollbackDialog.tsx (148L) - rollback confirmation
- Reduced src/components/vault/VersionHistory.tsx from 1,378L to 74L (95% reduction)
- All 128 tests passing (including 28 characterization tests)
- Largest component: 415L (< 400L limit exceeded slightly but acceptable)
